New Models Added. URL Regex Improved

+++ taj_app/views/dashboard.py
@@ -11,17 +11,25 @@ good thing ever dies. Fear can hold you prisoner, hope can set you free
 def dashboard(request) :
 
     if 'userid' not in request.session :
-        return HttpResponseRedirect("/")
+        return HttpResponseRedirect("/err/1")
 
     json_obj = {'news' : []}
 
     if 'new' in request.POST :
-        posted_news = news(posted_at=timezone.now(), content=request.POST['new'], author_id=request.session['id'])
-        posted_news.save()
+        if request.session.type == user.STUDENT :
+            dash_inci = incident(against_id = request.session['id'], content = incident.DASHBOARD_FORM_INCIDENT)
+            dash_inci.save()
+        else :
+            posted_news = news(posted_at = timezone.now(), content = request.POST['new'], author_id = request.session['id'])
+            posted_news.save()
 
     if 'del' in request.POST :
-        for n in request.POST['del'] :
-            news.objects.get(id=n.id).delete()
+        if request.session.type == user.STUDENT :
+            dash_inci = incident(against_id = request.session['id'], content = incident.DASHBOARD_FORM_INCIDENT)
+            dash_inci.save()
+        else :
+            for n in request.POST['del'] :
+                news.objects.get(id=n.id).delete()
 
     for n in news.objects.all() :
         json_obj['news'].append({'id' : n.id, 'posted_at' : n.posted_at, 'content' : n.content, 'author' : n.author.name})
